refactor: change folders, fix styles
