+++ src/features/auth/forgotPassword/index.ts
@@ -1,0 +1,1 @@
+export { ForgotPasswordForm } from './ui/ForgotPasswordForm'
