+++ src/features/auth/forgotPassword/ui/forgotPassword.module.scss
@@ -20,7 +20,6 @@
 }
 
 .text {
-  //margin-top: 0.5rem;
   margin-top: 7px;
   margin-bottom: 17px;
 
@@ -56,4 +55,10 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
+}
+
+.input {
+  & span {
+    position: static;
+  }
 }
