--- src/features/auth/forgotPassword/ui/index.ts
@@ -1,1 +1,0 @@
-export { ForgotPasswordForm } from './ForgotPasswordForm'
